t15-backtrace-d2.gdb: Break when out of memory
Put breakpoints on D's out-of-memory handlers
so that we can get more accurate stack traces.

Stack traces in exceptions may not be available
(because they need to be allocated in dynamic memory)
and the regular debugger stack traces may not be accurate
(because they may show a stack trace of the exception
being rethrown in a scope(exit) clause).