t15-backtrace-d2.gdb: Add a debug info hook
Add a hook that allows programs to inspect their internal state and
report any information that may be relevant in diagnosing the crash.

The return value could be anything that GDB can pretty-print: a C
string, a D string, or a struct, though support in GDB varies across
versions.

The function should not allocate memory as it may be run in an
out-of-memory situation.